Extracted ParticipantManager
